AI_UseCase/requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.30
chromadb>=0.4.24
sentence-transformers>=2.6.0
pypdf2>=3.0.1
pymupdf>=1.24.0
pydantic>=2.7.0
python-dotenv>=1.0.1
